[Bronze II] Title: 숫자, Time: 288 ms, Memory: 25248 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -11,19 +11,19 @@ public class Main {
 
         StringTokenizer st = new StringTokenizer(reader.readLine());
 
-        int n = Integer.parseInt(st.nextToken());
-        int m = Integer.parseInt(st.nextToken());
-        
+        long n = Long.parseLong(st.nextToken());
+        long m = Long.parseLong(st.nextToken());
+
         if (n == m) {
             System.out.println(0);
             return;
         }
-        
+
 
         System.out.println(Math.abs(n - m) - 1);
 
         StringBuilder builder = new StringBuilder();
-        for (int i = Math.min(n, m)+1; i < Math.max(n, m); i++) {
+        for (long i = Math.min(n, m)+1; i < Math.max(n, m); i++) {
             builder.append(i).append(" ");
         }
